// src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  MANAGER: 'MANAGER',
  HR_MANAGER: 'HR_MANAGER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const EmployeeStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  TERMINATED: 'TERMINATED'
} as const

export type EmployeeStatus = (typeof EmployeeStatus)[keyof typeof EmployeeStatus]


export const LeaveType = {
  SICK: 'SICK',
  CASUAL: 'CASUAL',
  VACATION: 'VACATION'
} as const

export type LeaveType = (typeof LeaveType)[keyof typeof LeaveType]


export const LeaveStatus = {
  PENDING: 'PENDING',
  APPROVED_BY_MANAGER: 'APPROVED_BY_MANAGER',
  APPROVED_BY_HR: 'APPROVED_BY_HR',
  REJECTED: 'REJECTED'
} as const

export type LeaveStatus = (typeof LeaveStatus)[keyof typeof LeaveStatus]


export const PayrollStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID'
} as const

export type PayrollStatus = (typeof PayrollStatus)[keyof typeof PayrollStatus]